requirements.txt
pandas>=1.3.0
numpy>=1.21.0
requests>=2.26.0
yfinance>=0.1.70
alpha-vantage>=2.3.1
beautifulsoup4>=4.10.0
lxml>=4.6.0
schedule>=1.1.0
python-dateutil>=2.8.0
matplotlib>=3.4.0
seaborn>=0.11.0
ta>=0.10.0
fredapi>=0.5.0
newsapi-python>=0.2.7
feedparser>=6.0.0
ollama>=0.6.1
tweepy>=4.12.0
flask>=2.3.0